Recipes

+++ Homie.API/DTOs/Food/UpdateFoodRequestBody.cs
@@ -6,11 +6,11 @@
         public string? Description { get; set; }
         public Guid? PlaceId { get; set; }
         public Guid? IngredientId { get; set; }
-        public float? IngredientAmount { get; set; }
+        public double? IngredientAmount { get; set; }
         public DateTime? ExpirationDate { get; set; }
         public DateTime? LastUpdated { get; set; }
 
-        public UpdateFoodRequestBody(string foodName, string? description, Guid? placeId, Guid? ingredientId, float? ingredientAmount, DateTime? lastUpdated, DateTime? expirationDate)
+        public UpdateFoodRequestBody(string foodName, string? description, Guid? placeId, Guid? ingredientId, double? ingredientAmount, DateTime? lastUpdated, DateTime? expirationDate)
         {
             FoodName = foodName;
             Description = description;
